# Speed up verify