Moved header sync patterns to private variables to speed up `verify`.  For VDIF, doing

```
header['sync_pattern'] == header._header_parser.defaults['sync_pattern']
```

takes 2.73 µs, while doing

```
header['sync_pattern'] == header._header_parser.defaults['sync_pattern'
```

takes 556 ns.  For Mark 4 the effect of using `np.all(header['sync_pattern'] == header._sync_pattern)` is less significant - 11 µs instead of 14.4 µs using `np.all(header['sync_pattern'] == header._header_parser.defaults['sync_pattern'])`.

Also removed frame verification when using `VDIFFrame.fromfile`, reducing 14.1 µs per frame read in.  About half of this (6.93 µs) is just used to check that the payload sample shape is the same as the header sample shape, so if we really wanted to we could just eliminate that instead of removing all verification.

Using the PR code to sequentially decode 2 billion complete samples (~500 MB) from an 8-thread VDIF file, we shave ~5 s / min (~8%) when compared to current master.

Addresses some of the comments in #133.